ConMan2: can now delete constraints from the UI. Debating whether I want to use a checkbox to modify button behavior or use a double-click like in the original script.
I'm leaning toward a checkbox to enable persistent behavior if the user wants it.

Fixed:
Clicking on an entry wouldn't always update the target list because the action was actually a click-drag into another entry. Changed the event to itemEnter to update when pressed cursor enters the item's list entry.