ConMan2: Switch!
This commit has been in progress for a while. This adds the entirety of the original switch functionality. This will require a double-check later to look for an offset update bug (unpredictable), but also for consistency between the switch functions. Mccabe complexity check is also complaining about switch_single() and __do_switch_on_off(), so should look for closures or splitting off a function or two.