Butter: regression: broke filtering all available curves if nothing selected
This came about with the new feature: filter selected keys. This indiscriminately gathered selected keyframes on the given curves without checking if a selection was made. Now it will gather all keys unless there are selections in the graph editor. This has been shown to not break when selecting one curve's keys out of several.